requirements.txt
fastapi
uvicorn
motor
pydantic
python-dotenv
python-multipart
cloudinary